Added support to implement decompression
Compress.py now stores the decoding key (dictionary) at the top of
the file rather than at the bottom.
Compress.py now stores the length of bitstring that was converted to
bytes so that it can re-convert the bytes back to bits while
decompressing

custom_nodes now supports an extra operation on all nodes -
clean_all()
This resets the static variables and attempts to delete all
pre-existing node objects so that the memory is free.
[Helpful for decompression]

custom_nodes had some unused methods removed.